+++ src/emailProvider.ts
@@ -1,70 +1,54 @@
-'use strict';
+"use strict";
 
-import * as vscode from 'vscode';
-import * as path from 'path';
-import * as fs from 'fs';
+import * as vscode from "vscode";
 
-import * as mailjet from 'node-mailjet';
+import * as mailjet from "node-mailjet";
 
-import helper from './helper';
+import helper from "./helper";
 
 export default class SendEmail {
 
     constructor(subscriptions: vscode.Disposable[]) {
         subscriptions.push(
-            vscode.commands.registerCommand('mjml.sendEmail', () => {
+            vscode.commands.registerCommand("mjml.sendEmail", () => {
                 this.sendEmail();
             })
         );
     }
 
-    private sendEmail() {
-        if (!(helper.isMJMLFile(vscode.window.activeTextEditor.document))) {
-            vscode.window.showWarningMessage('This is not a MJML document!');
-            return;
-        }
-
-        let content = helper.renderMJML(
-            vscode.window.activeTextEditor.document.getText(),
-            vscode.workspace.getConfiguration('mjml').minifyHtmlOutput,
-            vscode.workspace.getConfiguration('mjml').beautifyHtmlOutput
-        );
-
-        if (content) {
+    private sendEmail(): void {
+        helper.renderMJML((content: string) => {
             this.sendEmailWithMailjet(content);
-        }
-        else {
-            vscode.window.showErrorMessage('MJMLError: Failed to parse file ' + path.basename(vscode.window.activeTextEditor.document.uri.fsPath));
-        }
+        });
     }
 
-    private sendEmailWithMailjet(content: string) {
-        let defaultSubject = vscode.workspace.getConfiguration('mjml').mailSubject;
-        let defaultRecipients = vscode.workspace.getConfiguration('mjml').mailRecipients;
+    private sendEmailWithMailjet(content: string): void {
+        let defaultSubject: string = vscode.workspace.getConfiguration("mjml").mailSubject;
+        let defaultRecipients: string = vscode.workspace.getConfiguration("mjml").mailRecipients;
 
-        vscode.window.showInputBox({ placeHolder: 'Subject (' + defaultSubject + ')' }).then((subject) => {
-            vscode.window.showInputBox({ placeHolder: 'Recipients (' + defaultRecipients + ')' }).then((recipients) => {
+        vscode.window.showInputBox({ placeHolder: `Subject (${defaultSubject})` }).then((subject: string) => {
+            vscode.window.showInputBox({ placeHolder: `Recipients (${defaultRecipients})` }).then((recipients: string) => {
                 if (!subject) {
                     subject = defaultSubject;
                 }
 
-                recipients = (recipients ? recipients : defaultRecipients).replace(/\s/g, '').split(',').map((e) => {
-                    return {Email: e};
+                let recipientList: Array<{ Email: string }> = (recipients ? recipients : defaultRecipients).replace(/\s/g, "").split(",").map((emailAddress: string) => {
+                    return { Email: emailAddress };
                 });
 
                 mailjet.connect(
-                    vscode.workspace.getConfiguration('mjml').mailjetAPIKey,
-                    vscode.workspace.getConfiguration('mjml').mailjetAPISecret
-                ).post('send').request({
-                    FromEmail: vscode.workspace.getConfiguration('mjml').mailSender,
-                    FromName: vscode.workspace.getConfiguration('mjml').mailFromName,
+                    vscode.workspace.getConfiguration("mjml").mailjetAPIKey,
+                    vscode.workspace.getConfiguration("mjml").mailjetAPISecret
+                ).post("send").request({
+                    FromEmail: vscode.workspace.getConfiguration("mjml").mailSender,
+                    FromName: vscode.workspace.getConfiguration("mjml").mailFromName,
                     Subject: subject,
-                    Recipients: recipients,
-                    'Html-part': content
-                }).then(result => {
-                    vscode.window.showInformationMessage('Mail has been sent successfully.');
-                }).catch(err => {
-                    vscode.window.showErrorMessage('Something went wrong.');
+                    Recipients: recipientList,
+                    "Html-part": content
+                }).then((result: object) => {
+                    vscode.window.showInformationMessage("Mail has been sent successfully.");
+                }).catch((err: any) => {
+                    vscode.window.showErrorMessage("Something went wrong.");
                 });
             });
         });