+++ src/extension.ts
@@ -1,17 +1,18 @@
-'use strict';
+"use strict";
 
-import * as vscode from 'vscode';
-import * as childProcess from 'child_process';
-import * as path from 'path';
+import * as vscode from "vscode";
+import * as path from "path";
+import * as childProcess from "child_process";
 
-import * as phantomJS from 'phantomjs-prebuilt';
+import * as phantomJS from "phantomjs-prebuilt";
 
-import LintingProvider from './lintingProvider';
-import PreviewManager from './previewProvider';
-import ExportHTML from './exportProvider';
-import CopyHTML from './copyProvider';
-import Screenshot from './screenshotProvider';
-import SendEmail from './emailProvider';
+import LintingProvider from "./lintingProvider";
+import PreviewManager from "./previewProvider";
+import ExportHTML from "./exportProvider";
+import CopyHTML from "./copyProvider";
+import Screenshot from "./screenshotProvider";
+import SendEmail from "./emailProvider";
+import Template from "./templateProvider";
 
 let linter: LintingProvider;
 let previewManager: PreviewManager;
@@ -19,37 +20,38 @@ let exportHTML: ExportHTML;
 let copyHTML: CopyHTML;
 let screenshot: Screenshot;
 let sendEmail: SendEmail;
+let template: Template;
 
 export function activate(context: vscode.ExtensionContext) {
     // Gets a value indicating whether PhantomJS could be built
-    let phantomJSBuilt = null;
+    let phantomJSBuilt: any = undefined;
 
     // Rebuilding PhantomJS if required
     if (phantomJS.platform != process.platform) {
         try {
-            let env = process.env;
-            env['PHANTOMJS_PLATFORM'] = process.platform;
-            env['PHANTOMJS_ARCH'] = process.arch;
+            let env: NodeJS.ProcessEnv = process.env;
+            env["PHANTOMJS_PLATFORM"] = process.platform;
+            env["PHANTOMJS_ARCH"] = process.arch;
 
-            vscode.window.showInformationMessage('MJML needs to be rebuilt for your current platform. Please wait for the installation to finish...');
-            process.chdir(path.join(__dirname, '..'));
+            vscode.window.showInformationMessage("MJML needs to be rebuilt for your current platform. Please wait for the installation to finish...");
+            process.chdir(path.join(__dirname, ".."));
 
-            childProcess.exec('npm --strict-ssl false rebuild phantomjs-prebuilt', {
+            childProcess.exec("npm --strict-ssl false rebuild phantomjs-prebuilt", {
                 env: env
-            }, (error, stdout, stderr) => {
+            }, (error: Error, stdout: string, stderr: string) => {
                 if (!error && !stderr) {
                     phantomJSBuilt = true;
-                    vscode.window.showInformationMessage('MJML\'s been updated. Please restart VSCode in order to continue using MJML.');
+                    vscode.window.showInformationMessage("MJML's been updated. Please restart VSCode in order to continue using MJML.");
                 }
                 else {
-                    vscode.window.showErrorMessage('MJML couldn\'t build the propper version of PhantomJS. Restart VSCode in order to try it again.');
+                    vscode.window.showErrorMessage("MJML couldn't build the propper version of PhantomJS. Restart VSCode in order to try it again.");
                 }
 
                 screenshot = new Screenshot(context, process.platform, phantomJS.platform, phantomJSBuilt);
             });
         }
         catch (e) {
-            vscode.window.showErrorMessage('MJML couldn\'t build the propper version of PhantomJS. Restart VSCode in order to try it again.');
+            vscode.window.showErrorMessage("MJML couldn't build the propper version of PhantomJS. Restart VSCode in order to try it again.");
             phantomJSBuilt = false;
         }
     }
@@ -57,7 +59,7 @@ export function activate(context: vscode.ExtensionContext) {
         screenshot = new Screenshot(context, process.platform, phantomJS.platform, phantomJSBuilt);
     }
 
-    if (vscode.workspace.getConfiguration('mjml').lintEnable) {
+    if (vscode.workspace.getConfiguration("mjml").lintEnable) {
         linter = new LintingProvider(context.subscriptions);
     }
 
@@ -65,6 +67,7 @@ export function activate(context: vscode.ExtensionContext) {
     exportHTML = new ExportHTML(context.subscriptions);
     copyHTML = new CopyHTML(context.subscriptions);
     sendEmail = new SendEmail(context.subscriptions);
+    template = new Template(context.subscriptions);
 }
 
 export function deactivate() {