+++ src/helper.ts
@@ -1,17 +1,45 @@
-'use strict';
+"use strict";
 
-import * as vscode from 'vscode';
-import * as path from 'path';
+import * as vscode from "vscode";
+import * as path from "path";
 
-import * as fileUrl from 'file-url';
-import { mjml2html } from 'mjml';
+import fileUrl = require("file-url");
+import { mjml2html } from "mjml";
 
 export default class Helper {
 
-    static renderMJML(mjml: string, minify: boolean, beautify: boolean): string {
+    static renderMJML(cb: (content: string) => any, fixLinks?: boolean, minify?: boolean, beautify?: boolean): any {
+        if (!(this.isMJMLFile(vscode.window.activeTextEditor.document))) {
+            vscode.window.showWarningMessage("This is not a MJML document!");
+            return;
+        }
+
+        let content: string = this.mjml2html(
+            vscode.window.activeTextEditor.document.getText(),
+            minify != undefined ? minify : vscode.workspace.getConfiguration("mjml").minifyHtmlOutput,
+            beautify != undefined ? beautify : vscode.workspace.getConfiguration("mjml").beautifyHtmlOutput,
+        );
+
+        if (content) {
+            if (fixLinks != undefined && fixLinks) {
+                content = this.fixLinks(content);
+            }
+
+            return cb(content);
+        }
+        else {
+            vscode.window.showErrorMessage(`MJMLError: Failed to parse file ${path.basename(vscode.window.activeTextEditor.document.uri.fsPath)}`);
+        }
+    }
+
+    static isMJMLFile(document: vscode.TextDocument): boolean {
+        return document.languageId === "mjml" && document.uri.scheme !== "mjml-preview";
+    }
+
+    static mjml2html(mjml: string, minify: boolean, beautify: boolean): string {
         try {
-            let html = mjml2html(mjml, {
-                level: 'skip',
+            let html: any = mjml2html(mjml, {
+                level: "skip",
                 minify: minify,
                 beautify: beautify,
                 filePath: vscode.window.activeTextEditor.document.uri.fsPath,
@@ -40,13 +68,9 @@ export default class Helper {
         return text.replace(
             new RegExp(/((?:src|href|url)(?:=|\()(?:[\'\"]|))((?!http|\\|"|#).+?)([\'\"]|\))/, "gmi"),
             (subString: string, p1: string, p2: string, p3: string): string => {
-                return [p1, fileUrl(path.join(path.dirname(vscode.window.activeTextEditor.document.uri.fsPath), p2)), p3].join('');
+                return [p1, fileUrl(path.join(path.dirname(vscode.window.activeTextEditor.document.uri.fsPath), p2)), p3].join("");
             }
         );
-    }
-
-    static isMJMLFile(document: vscode.TextDocument): boolean {
-        return document.languageId === 'mjml' && document.uri.scheme !== 'mjml-preview';
     }
 
 }